requirements.txt
python-dotenv
langchain
openai
streamlit
faiss-cpu
huggingface-hub
PyPDF2
InstructorEmbedding
sentence-transformers
tiktoken
qdrant_client